requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.7
django-environ==0.4.5
django-filter==2.4.0
django-heroku==0.3.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
importlib-metadata==3.10.0
Markdown==3.3.4
psycopg2==2.8.6
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.4
whitenoise==5.2.0
zipp==3.4.1
